First step migrating further into clean arch

+++ backend/src/application/use_cases/author_ops/delete_author.py
@@ -1,7 +1,10 @@
-from sqlalchemy.orm import Session
+from backend.src.application.interfaces.author_repository import \
+    AuthorRepository
+from backend.src.domain.exceptions.author_exceptions import AuthorNotFound
 
-from backend.src.application.use_cases import author_ops
-from backend.src.infrastructure.persistence.author_repository_impl import AuthorRepository
 
 def delete_author(author_repo: AuthorRepository, author_id: int) -> bool:
+    if not author_repo.get_by_id(author_id=author_id):
+        raise AuthorNotFound("There exists no author of such ID")
+    
     return author_repo.delete(author_id)
